requirements.txt
python-telegram-bot[webhooks]==21.4
python-dotenv==1.0.1
aiohttp==3.9.5
pandas==2.2.2
openpyxl==3.1.2
